+++ webapp/src/features/charts/components/radar-benef-control.tsx
@@ -1,60 +1,7 @@
 import type { FC } from "react";
-import {
-  PolarAngleAxis,
-  PolarGrid,
-  PolarRadiusAxis,
-  Radar,
-  RadarChart,
-} from "recharts";
+import { PolarGrid, PolarRadiusAxis, Radar, RadarChart } from "recharts";
 
 import { useTranslation } from "@shared/i18n";
-
-const wrapPolarLabel = (label: string, maxChars = 14) => {
-  const words = label.split(" ");
-  const lines: string[] = [];
-  const nbLines = Math.ceil(label.length / maxChars);
-  const avgLineLength = label.length / nbLines;
-  let currentLine = "";
-
-  for (const word of words) {
-    const next = currentLine ? `${currentLine} ${word}` : word;
-    if (next.length < avgLineLength || lines.length === nbLines - 1) {
-      currentLine = next;
-    } else {
-      if (currentLine) lines.push(currentLine);
-      currentLine = word;
-    }
-  }
-
-  if (currentLine) lines.push(currentLine);
-  return lines;
-};
-
-const renderPolarAngleTick = ({ payload, x, y, textAnchor }: any) => {
-  const label = String(payload?.value ?? "");
-  const lines = wrapPolarLabel(label);
-
-  return (
-    <text
-      fill="#9c9798"
-      fontSize={12}
-      textAnchor={textAnchor}
-      x={x}
-      y={y}
-    >
-      {lines.map((line, index) => (
-        <tspan
-          dy={index === 0 ? 0 : 16}
-          // biome-ignore lint/suspicious/noArrayIndexKey: <don't want to enforce id>
-          key={index}
-          x={x}
-        >
-          {line}
-        </tspan>
-      ))}
-    </text>
-  );
-};
 
 import { Card, CardContent, CardDescription, CardHeader } from "@ui/card";
 import {
@@ -67,6 +14,7 @@ import {
 } from "@ui/chart";
 
 import { RADAR_CONFIG } from "../constants";
+import { PolarAngleAxisMultiline } from "./polar-angle-axis-multi";
 
 type ChartRadarWithBenefAndControlProps = {
   title: string;
@@ -112,10 +60,7 @@ export const ChartRadarWithBenefAndControl: FC<
               domain={[0, 10]}
               tickCount={6}
             />
-            <PolarAngleAxis
-              dataKey="indicator"
-              tick={renderPolarAngleTick}
-            />
+            <PolarAngleAxisMultiline dataKey="indicator" />
             <PolarGrid radialLines />
 
             <Radar
