+++ webapp/src/features/charts/socio-eco/chart-wood-energy-needs.tsx
@@ -71,7 +71,7 @@ export const ChartWoodEnergyNeeds: FC<PieChartProps> = ({ data }) => {
             data={chartData}
             dataKey="value"
           />
-          <Tooltip defaultIndex={3} />
+          <Tooltip />
         </PieChart>
       </ResponsiveContainer>
     </div>
