+++ webapp/src/shared/i18n/translations/fr/translations.json
@@ -131,12 +131,12 @@
             "coal": "Charbon",
             "collectedWood": "Bois collecté",
             "gas": "Gaz",
-            "legend": "% ménages utilisant la ressource",
+            "legend": "ménages utilisant la ressource (en %):",
             "organicWaste": "Déchets végétaux",
             "other": "Autres",
             "title": "Sources d’énergie utilisées"
           },
-          "lumberNeeds": {
+          "timberNeeds": {
             "difficultToMeet": "Difficile à satisfaire",
             "dontKnow": "Ne sait pas",
             "easyToMeet": "Facile à satisfaire",
