+++ webapp/src/shared/ui/chart.tsx
@@ -241,7 +241,10 @@ const ChartTooltipContent = React.forwardRef<
                         </div>
                         {item.value && (
                           <span className="font-mono font-medium tabular-nums text-foreground">
-                            {item.value.toLocaleString()}
+                            {
+                              /* Force a space between item label and value*/ "\xa0" +
+                                item.value.toLocaleString()
+                            }
                           </span>
                         )}
                       </div>
